# Conflicts:
#	src/bincode.rs

+++ src/bincode.rs
@@ -4,11 +4,17 @@ use bincode::enc::Encoder;
 use bincode::error::{DecodeError, EncodeError};
 use bincode::{BorrowDecode, Decode, Encode};
 
+// Since Slab is stored as a map of usize to T, its implementations are quite similar to those of HashMap<usize, T>:
+// https://github.com/bincode-org/bincode/blob/v2.0.0-rc.2/src/features/impl_std.rs#L409-L469.
+
 impl<T: Encode> Encode for Slab<T> {
     fn encode<E: Encoder>(&self, encoder: &mut E) -> Result<(), EncodeError> {
-        (self.len() as u64).encode(encoder)?;
+        // https://docs.rs/bincode/2.0.0-rc.2/bincode/config/struct.Configuration.html#method.with_fixed_int_encoding
+        // With fixed integer encoding, usize is always encoded as a u64,
+        // so there's no need to worry about varying sizes of usize.
+        self.len().encode(encoder)?;
         for (key, value) in self {
-            (key as u64).encode(encoder)?;
+            key.encode(encoder)?;
             value.encode(encoder)?;
         }
         Ok(())
@@ -17,14 +23,15 @@ impl<T: Encode> Encode for Slab<T> {
 
 impl<T: Decode> Decode for Slab<T> {
     fn decode<D: Decoder>(decoder: &mut D) -> Result<Self, DecodeError> {
-        let len = u64::decode(decoder)? as usize;
-        decoder.claim_container_read::<(u64, T)>(len)?;
+        let len = usize::decode(decoder)?;
+        decoder.claim_container_read::<(usize, T)>(len)?;
 
         let mut builder = Builder::with_capacity(len);
         for _ in 0..len {
-            decoder.unclaim_bytes_read(core::mem::size_of::<(u64, T)>());
+            // See the documentation on `unclaim_bytes_read` as to why we're doing this here
+            decoder.unclaim_bytes_read(core::mem::size_of::<(usize, T)>());
 
-            let key = u64::decode(decoder)? as usize;
+            let key = usize::decode(decoder)?;
             let value = T::decode(decoder)?;
             builder.pair(key, value);
         }
@@ -34,14 +41,15 @@ impl<T: Decode> Decode for Slab<T> {
 
 impl<'de, T: BorrowDecode<'de>> BorrowDecode<'de> for Slab<T> {
     fn borrow_decode<D: BorrowDecoder<'de>>(decoder: &mut D) -> Result<Self, DecodeError> {
-        let len = u64::decode(decoder)? as usize;
-        decoder.claim_container_read::<(u64, T)>(len)?;
+        let len = usize::decode(decoder)?;
+        decoder.claim_container_read::<(usize, T)>(len)?;
 
         let mut builder = Builder::with_capacity(len);
         for _ in 0..len {
-            decoder.unclaim_bytes_read(core::mem::size_of::<(u64, T)>());
+            // See the documentation on `unclaim_bytes_read` as to why we're doing this here
+            decoder.unclaim_bytes_read(core::mem::size_of::<(usize, T)>());
 
-            let key = u64::borrow_decode(decoder)? as usize;
+            let key = usize::borrow_decode(decoder)?;
             let value = T::borrow_decode(decoder)?;
             builder.pair(key, value);
         }
